requirements.txt
discord.py
requests
texttable
tabulate
numpy
BeautifulSoup4 
